KEYCLOAK-4187 Extract InfinispanKeycloakTransaction from InfinispanUserSessionProvider
